Many AI changes
Fixed an issue where weighted randomization caused AI-units to reposition over and over until all their TUs are spent leaving them with no chance to reaction-fire. The units will instead mark themselves as wanting to end their turn after the first non-attacking and non-peeking-move and only wake up again if any of their team-members finds something interesting.

Weighted randomization was merged with intelligence-scaling. The old intelligence-scaling that made the AI do random-moves was way too extreme in making the game easier, which was against the design-philosophy of brutal-AI. So instead intelligence now impacts the roll-range for the random-score-modifier of the previous weigthed randomization. The lower-boundary is 0.2 times intelligence. The upper boundary remains 1.0. Attack-moves are now also subject to the intelligence-roll.

Units that have made a move with the intent of getting into an attacking-position will no longer perform a second iteration of the movement-logic to check their time-unit-preservation in order to prevent indecisive behavior caused by application of a random score-modifier.

The way aggressiveness works is now way more gradual and doesn't alter the behaviour in extreme steps like the previous integration.
The way it works now is that an aggressiveness-based factor is applied to the "will to be closer to the enemy". This factor gradually shifts the behavior from careful to aggressive.

Baseline-aggressiveness can now be set with two numerical values from 1-9. One acting as numerator and one as denominator. So aggressiveness can be set up from 1/9th to 9.
The meaning of this value roughly is: "How strong do I feel compared to one unit of the enemy."

The aggression-mode-setting now has only the option to multiply unit-aggression with the baseline-aggressiveness instead of replacing it. This way the baseline-aggressiveness can be adjusted to the aggression-values used within the mod to produce the desired results. For example if 2 is supposed to be the avarage balanced behavior the baseline-aggressiveness should be set to 1 over 2.

The aggressiveness is now always impacted by both morale as well as a rough estimate of each team's total combat-power. This way the AI will situationally adapt to what's happening and decide if it shall be more defensive or more aggressive compared to the baseline.

An aggressiveness of 0 can only be reached by units with 0 aggressivness when aggression is inherited. This is the only value at which the AI will deliberately disengage. Any value above 0 will at the very most be as defensive as the previous aggressiveness 1 behaved. That means they will still move up to the closest tile they deem to be save based on their scouting-information.

An aggressiveness of "1" is now a threshold for two behavioral shifts: Below that threshold there is an increasing score-bonus for being indoors. The reciprocal of the aggressivness is multiplied with the score of a tile, when this tile has a roof above.

The other shift is that above a value of 1, the AI no longer will prefer reserving time-units for breaking line of sight over performing an additional attack.

The previous completely reckless behavior now is only possible with units that have the Leeroy-flag and at aggression-mode 2. Otherwise high aggressiveness-values will not put a lot of emphasis of taking cover but will still make use of existing cover if it is opportune while closing in.

Also non-Leeroy-units that know they have been seen during their turn will always want to take cover regardless of how high their aggressivness is.

+++ src/Savegame/BattleUnit.h
@@ -815,7 +815,7 @@ public:
 	/// Get the unit leeroyJenkins flag
 	bool isLeeroyJenkins(bool ignoreBrutal = false) const;
 	/// Get the unit's aggression-flag
-	int getAggressiveness() const;
+	float getAggressiveness() const;
 	/// Get the unit's intelligence
 	int getBrutalIntelligence() const;
 	/// Gets the spotter score. This is the number of turns sniper AI units can use spotting info from this unit.
